Show Persian time in Appointment List

+++ composeApp/src/commonMain/kotlin/com/sample/turnapp/feature/appointment/domain/SaveAppointmentParam.kt
@@ -3,8 +3,8 @@ package com.sample.turnapp.feature.appointment.domain
 data class SaveAppointmentParam(
     val id: Int? = null,
     val personId: Int,
-    val startTime: Double,
-    val endTime: Double,
+    val startTime: Long,
+    val endTime: Long,
     val title: String,
     val description: String? = null
 )
